+++ sample/src/commonMain/kotlin/com/kez/picker/sample/ui/navigation/Screen.kt
@@ -5,6 +5,7 @@ sealed class Screen(val route: String) {
     object WheelPicker : Screen("wheel_picker")
     object Integrated : Screen("integrated")
     object TimePicker : Screen("time_picker")
+    object DurationPicker : Screen("duration_picker")
     object YearMonthPicker : Screen("year_month_picker")
     object DatePicker : Screen("date_picker")
     object DateRangePicker : Screen("date_range_picker")